Created Alpha branch and added info to contributing.md
I have created a new branch called Alpha that I think we should be using as a pre-master branch from now on. So all pull requests are merged into the alpha and once the alpha is stable we can merge that into the master branch. As a way to keep the master branch as clean as possible.